+++ inc/WPCLI/CLI.php
@@ -5,7 +5,6 @@ namespace ChubesDocs\WPCLI;
 use ChubesDocs\WPCLI\Commands\ProjectEnsureCommand;
 use ChubesDocs\WPCLI\Commands\ProjectTreeCommand;
 use ChubesDocs\WPCLI\Commands\DocsSyncCommand;
-use ChubesDocs\WPCLI\Commands\TaxonomyMigrateCommand;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -16,6 +15,5 @@ class CLI {
 		\WP_CLI::add_command( 'chubes project ensure', [ ProjectEnsureCommand::class, 'run' ] );
 		\WP_CLI::add_command( 'chubes project tree', [ ProjectTreeCommand::class, 'run' ] );
 		\WP_CLI::add_command( 'chubes docs sync', [ DocsSyncCommand::class, 'run' ] );
-		\WP_CLI::add_command( 'chubes taxonomy migrate', [ TaxonomyMigrateCommand::class, 'run' ] );
 	}
 }